refactor(ui): remove duplicate appearance settings
