+++ ui/src/lib/adminNavigation.test.ts
@@ -31,7 +31,6 @@ describe('isOwnerOnlyPath', () => {
   });
 
   it('keeps personal preferences, replies, and access readable', () => {
-    expect(isOwnerOnlyPath('/settings/appearance')).toBe(false);
     expect(isOwnerOnlyPath('/settings/account')).toBe(false);
     expect(isOwnerOnlyPath('/settings/replies')).toBe(false);
     expect(isOwnerOnlyPath('/settings/access')).toBe(false);
@@ -52,14 +51,19 @@ describe('settings landing memory', () => {
     expect(settingsLandingPath(true)).toBe('/settings/backends/claude');
   });
 
-  it('falls back to Appearance for members when an owner-only section was remembered', () => {
+  it('falls back to Account for members when an owner-only section was remembered', () => {
     window.localStorage.setItem(SETTINGS_LAST_PATH_KEY, '/settings/service');
-    expect(settingsLandingPath(false)).toBe('/settings/appearance');
+    expect(settingsLandingPath(false)).toBe('/settings/account');
   });
 
   it('does not remember transitional platform scope pages as the Settings landing page', () => {
     rememberSettingsPath('/settings/platforms/groups');
-    expect(settingsLandingPath(true)).toBe('/settings/appearance');
+    expect(settingsLandingPath(true)).toBe('/settings/account');
+  });
+
+  it('retires a remembered Appearance destination', () => {
+    window.localStorage.setItem(SETTINGS_LAST_PATH_KEY, '/settings/appearance');
+    expect(settingsLandingPath(true)).toBe('/settings/account');
   });
 });
