+++ ui/src/lib/adminNavigation.ts
@@ -49,6 +49,7 @@ export const isLocalOnlyMessagingField = (field: string): boolean =>
 
 const isValidSettingsPath = (pathname: string): boolean =>
   pathname.startsWith('/settings/') &&
+  pathname !== '/settings/appearance' &&
   !pathname.startsWith('/settings/platforms/groups') &&
   !pathname.startsWith('/settings/platforms/users');
 
@@ -70,7 +71,7 @@ export const settingsLandingPath = (canManageInstance: boolean): string => {
       // Ignore unavailable storage.
     }
   }
-  if (!remembered || !isValidSettingsPath(remembered)) return '/settings/appearance';
-  if (!canManageInstance && isOwnerOnlyPath(remembered)) return '/settings/appearance';
+  if (!remembered || !isValidSettingsPath(remembered)) return '/settings/account';
+  if (!canManageInstance && isOwnerOnlyPath(remembered)) return '/settings/account';
   return remembered;
 };
